backend/requirements.txt
fastapi==0.68.2
uvicorn==0.15.0
python-dotenv==0.19.0
openai==1.0.0
beautifulsoup4==4.12.0
aiohttp==3.9.0
lxml==4.9.0
requests==2.31.0
crawl4ai==0.1.0
bs4==0.0.1 